chore: rename amd dir
