+++ lunal-attestation/src/lib.rs
@@ -4,7 +4,7 @@ pub mod attestation;
 // #[cfg(target_arch = "wasm32")]
 // pub mod wasm;
 
-pub mod amd_azure;
+pub mod amd;
 pub mod pcs_client;
 pub mod utils;
 pub mod verify;
